Add a generic component (SearchQueryContainer) to be reused into the different pages that uses search and facets queries.
