+++ .eslintrc
@@ -1,0 +1,1 @@
+/home/estacio/.config/yarn/global/node_modules/vtex/.eslintrc
