+++ react/BrandPage.js
@@ -2,7 +2,19 @@ import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { ExtensionPoint } from 'render'
 
+import SearchQueryContainer from './components/SearchQueryContainer'
+import { searchQueryPropTypes } from './constants/propTypes'
+import { SearchQueryContext } from './constants/searchContext'
+
 export default class BrandPage extends Component {
+  static propTypes = {
+    params: PropTypes.shape({
+      /** Brand name */
+      brand: PropTypes.string.isRequired,
+    }),
+    ...searchQueryPropTypes,
+  }
+
   static contextTypes = {
     prefetchPage: PropTypes.func,
   }
@@ -13,7 +25,11 @@ export default class BrandPage extends Component {
 
   render() {
     return (
-      <ExtensionPoint id="container" />
+      <SearchQueryContainer {...this.props}>
+        <SearchQueryContext.Consumer>
+          {contextProps => <ExtensionPoint id="container" {...contextProps} />}
+        </SearchQueryContext.Consumer>
+      </SearchQueryContainer>
     )
   }
 }
