+++ react/constants/searchContext.js
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const SearchQueryContext = React.createContext()
